2024/12/01
Changed all resolutions in each screen to 1000 x 667 to give more space for stuff.
Added the main menu buttons and changed the color since it was too hard to do anything with just white. Also had to scale to SplashScreen.
Replaced Extra button with Tutorial.
Added a sound when you click any of those buttons. There is also a main menu sound.
Added a sound effect each time you throw the ball.
Added the tries.(Like the guy got X attempts to make the shot)
Added the hoop.
Added a deadzone which needs to be re-implemented but its the box thing that we thought about to block the ball.
Added text showing how many more tries they got.
Changed the velocity of the ball from 1200 to 1400 to as it couldnt reach the hoop.